[Rust] Unified multi-pool radix tree port (plan M2/1e)
Port the UnifiedTreeCore tree semantics (the multi-pool
unified_radix_cache.py engine) into sglang-radix: one node arena with
per-component device + host values (FULL/SWA/Mamba/C128), per-ct
intrusive LRUs, namespaced child keys under a single root, the stepped
resumable insert walk with split/unevict/overlap barriers, the
multi-validator match with device/host boundaries + host-hit lengths,
inc/dec lock walks, stepwise device eviction (FULL leaf heap, SWA/Mamba
LRU cursors) with demote/delete/deferred write-back backup,
drop_subtree_no_host, host-tier insert_host + drive_host_eviction with
write_back duplicate-host reclaim, write-through/load-back pending
commits, KV event log, sanity_check and node dumps. No torch crosses
the boundary: values are pool-index lists, allocator calls return as
actions the controller drains.

Exposure + Python side:
- UnifiedRadixTree pyclass in sglang-scheduler pybind (CT_*/PHASE_*
  module constants).
- UnifiedTreeCoreRust wrapper behind the UnifiedTreeCoreInterface
  (tree_core_rust.py) + 'rust' tree-core backend in
  tree_core_registry.py (env SGLANG_UNIFIED_RADIX_TREE_CORE_BACKEND).
  Components keep only facade-level hooks under the Rust backend.

Verification:
- 56 library + 16 integration Rust tests (torch-free).
- 15 criterion benches (unified_bench: insert/match/evict drains/
  hicache/lock/canary/clone over the 256-agent shape).
- test_rust_unified_radix_parity.py (CPU CI, differential vs the
  unmodified Python UnifiedTreeCore with the real FullComponent,
  FULL-only, shared op script; write-back round trip + insert_host +
  host-pressure reclaim in a second script).
- Also: Clone derives on the unified tree (bench clone-per-iteration
  convention) and the routine clippy fixes they surfaced.

Host-gated remainder (as with M2/1b-1d): the parity test runs in CPU
CI (no torch on the dev host); e2e A/B stays on the target host.

+++ rust/sglang-radix/benches/radix_bench.rs
@@ -361,7 +361,7 @@ fn bench_mamba_evict(c: &mut Criterion) {
         });
     }
     // 1 shared internal state + 255 tail states.
-    let m = (256usize as f64 * 0.1) as usize;
+    let m = (256_f64 * 0.1) as usize;
     group.bench_with_input("agent-256_mamba-10pct", &m, |b, &m| {
         b.iter(|| {
             let mut t = tree.clone();
